fixed some font sizing
